bwrap-oci: move more functions to the util module
